Move `AnimatedVisibility`
