+++ packages/fuselage/src/components/Box/Position/PositionAnimated.tsx
@@ -1,7 +1,7 @@
 import React, { ComponentProps } from 'react';
 
 import Position from '.';
-import AnimatedVisibility from '../AnimatedVisibility';
+import AnimatedVisibility from '../../AnimatedVisibility';
 
 type PositionAnimatedProps = {
   visible?: 'hidden' | 'visible' | 'hiding' | 'unhiding';
