+++ packages/fuselage/src/components/Box/index.js
@@ -104,5 +104,4 @@ export const Box = memo(
 
 Box.displayName = 'Box';
 
-export { default as AnimatedVisibility } from './AnimatedVisibility';
 export { default as Position, PositionAnimated } from './Position';
